+++ tartiflette/engine.py
@@ -1,8 +1,10 @@
-from typing import Dict, Any
+from typing import Any, Dict
+
+from tartiflette.executors.basic import execute as basic_execute
 from tartiflette.parser import TartifletteRequestParser
 from tartiflette.schema.registry import SchemaRegistry
 from tartiflette.schema.bakery import SchemaBakery
-from tartiflette.executors.basic import execute as basic_execute
+from tartiflette.types.exceptions.tartiflette import GraphQLError
 
 
 class Engine:
@@ -26,7 +28,7 @@ class Engine:
         query: str,
         context: Dict[str, Any] = None,
         variables: Dict[str, Any] = None,
-    ) -> str:
+    ) -> dict:
         """
         Parse and execute a GraphQL Request (as string).
         :param query: The GraphQL request / query as UTF8-encoded string
@@ -34,9 +36,13 @@ class Engine:
         :param variables: The variables used in the GraphQL request
         :return: a GraphQL Response (as dict)
         """
-        return await basic_execute(
-            self._parser.parse_and_tartify(
+        try:
+            root_nodes = self._parser.parse_and_tartify(
                 self._schema, query, variables=variables
-            ),
-            request_ctx=context,
-        )
+            )
+        except GraphQLError as e:
+            return {"data": None, "errors": [e.coerce_value()]}
+        except Exception:  # pylint: disable=broad-except
+            gql_error = GraphQLError("Server encountered an error.")
+            return {"data": None, "errors": [gql_error.coerce_value()]}
+        return await basic_execute(root_nodes, request_ctx=context)